Added tpu mode

+++ GanTrainer/GanTrainer.py
@@ -88,8 +88,8 @@ class GanTrainer:
 
         train_size = self.X_train.shape[0]
 
-        valid = np.ones((batch_size, 1))
-        fake = np.zeros((batch_size, 1))
+        valid = np.ones((batch_size, 1)).astype(np.float16)
+        fake = np.zeros((batch_size, 1)).astype(np.float16)
 
         batch_learned = 0
         epoch = 0
